pyproject.toml
[project]
name = "mailroom"
version = "0.1.0"
description = "Email triage automation for Fastmail using JMAP"
readme = "README.md"
authors = [
    { name = "Flo Kempenich", email = "flo@kempenich.dev" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "pydantic-settings",
    "structlog",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-httpx",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
